Extract logscan progress/incomplete-run analysis into modules/logscan_progress.py
Fourth sub-phase of the logscan/trends breakup, and the heaviest in
monkeypatch density so far. Moves progress-config loading, the
per-library progress-snapshot table builder shared by completed and
incomplete runs, the incomplete-run resume analyzer (built on top of
the recovery-suggestion layer from sub-phase 2), and the cache-backed
incomplete-run lookups used by the trends dashboard and resume hint.

Five functions in this set are individually monkeypatched by tests as
dependencies of other functions moving together in this phase
(_load_progress_config, _analyze_incomplete_log_for_resume,
_get_logscan_incomplete_runs, _get_logscan_incomplete_run,
_get_incomplete_resume_runs) - reached via the established deferred
`import quickstart` bridge at each of their 7 internal call sites.

Missed one on the first pass: _load_logscan_ingest_cache (from
modules/logscan_cache.py, sub-phase 1) is also heavily monkeypatched,
and the three functions here that read the ingest cache
(get_logscan_incomplete_runs, get_logscan_incomplete_run,
get_incomplete_resume_runs) were calling the directly-imported
function instead of bridging through quickstart - broke 3 tests that
monkeypatch qs_module._load_logscan_ingest_cache and expect these
functions to observe it. Fixed by routing those three call sites
through the same bridge.
